+++ src/main.cpp
@@ -1,3 +1,4 @@
+#include <atomic>
 #include <iostream>
 #include <signal.h>
 
@@ -6,17 +7,16 @@
 #include "http/server-ipv6.hpp"
 #include "logs/logger.hpp"
 #include "http/version_checker.hpp"
+#include "util/string_tools.hpp"
+
+// Global termination flag (atomic)
+std::atomic<bool> isExiting{false};
 
 std::vector<std::shared_ptr<http::Server>> serversVec;
 
 /******************** SIGNAL HANDLERS & CLEANUP ********************/
 
-void cleanExit(); // Fwd declaration
-void catchSig(int s) {
-    std::cout << "\nIntercepted exit signal " << s << ", closing...\n";
-    cleanExit();
-    exit(0);
-}
+void catchSig(int) { isExiting.store(true); }
 
 void initSigHandler() {
     #ifdef _WIN32
@@ -38,10 +38,6 @@ void initSigHandler() {
 }
 
 void cleanExit() {
-    for (auto pServer : serversVec) {
-        pServer->kill(); // Kill the server
-    }
-
     #ifdef _WIN32
         WSACleanup();
     #endif
@@ -53,21 +49,65 @@ void cleanExit() {
     conf::cleanupConfig();
 }
 
+void awaitExitCin() {
+    // Otherwise, read from cin
+    std::cout << "> ";
+    std::string buf;
+    std::getline(std::cin, buf);
+    trimString(buf); strToUpper(buf);
+
+    // Clean exit
+    if (buf == "EXIT")
+        isExiting.store(true);
+}
+
 /******************** WELCOME BANNER METHODS ********************/
 
 void printCenteredVersion() {
     const int leftPadding = (34 - conf::VERSION.length()) / 2;
     const int rightPadding = 34 - conf::VERSION.length() - leftPadding;
 
-    std::cout << '|' << std::string(leftPadding, ' ') << conf::VERSION << std::string(rightPadding, ' ') << '|' << '\n';
+    #ifdef _WIN32
+        std::wcout << L'\u00B3';
+        std::cout << std::string(leftPadding, ' ') << conf::VERSION << std::string(rightPadding, ' ');
+        std::wcout << L'\u00B3';
+        std::cout << std::endl;
+    #else
+        std::cout << "\u2502" << std::string(leftPadding, ' ') << conf::VERSION << std::string(rightPadding, ' ') << "\u2502" << std::endl;
+    #endif
 }
 
 void printWelcomeBanner() {
-    std::cout << "------------------------------------\n";
-    printCenteredVersion();
-    std::cout << "|           ...........            |\n"
-                 "|         Ctrl+C to close.         |\n"
-                 "------------------------------------\n";
+    /**
+     * Prints the following message:
+     * 
+     * ┌──────────────────────────────────┐
+     * │          Mercury vX.X.X          │
+     * │           ════════════           │
+     * │         "Exit" to close.         │
+     * └──────────────────────────────────┘
+     */
+    #ifdef _WIN32
+        std::wcout << L'\u00DA';
+        for (int i = 0; i < 34; ++i) std::wcout << L'\u00C4';
+        std::wcout << L'\u00BF' << std::endl;
+        printCenteredVersion();
+        std::wcout << L"\u00B3           \u00CD\u00CD\u00CD\u00CD\u00CD\u00CD\u00CD\u00CD\u00CD\u00CD\u00CD\u00CD           \u00B3\n"
+                      L"\u00B3         \"Exit\" to close.         \u00B3\n"
+                      L"\u00C0";
+        for (int i = 0; i < 34; ++i) std::wcout << L'\u00C4';
+        std::wcout << L'\u00D9' << std::endl;
+    #else
+        std::cout << "\u250C";
+        for (int i = 0; i < 34; ++i) std::cout << "\u2500";
+        std::cout << "\u2510" << std::endl;
+        printCenteredVersion();
+        std::cout << "\u2502           \u2550\u2550\u2550\u2550\u2550\u2550\u2550\u2550\u2550\u2550\u2550\u2550           \u2502\n"
+                      "\u2502         \"Exit\" to close.         \u2502\n"
+                      "\u2514";
+        for (int i = 0; i < 34; ++i) std::cout << "\u2500";
+        std::cout << "\u2518" << std::endl;
+    #endif
     ACCESS_LOG << conf::VERSION << " started successfully." << std::endl;
 }
 
@@ -75,7 +115,7 @@ void printWelcomeBanner() {
 
 int main() {
     // Initialize Winsock API
-    #if defined(WIN32) || defined(_WIN32) || defined(__WIN32__) || defined(__NT__)
+    #ifdef _WIN32
         WSADATA wsa;
         if (WSAStartup(MAKEWORD(2, 2), &wsa) != 0) {
             std::cerr << "Failed to initialize Winsock API.\n";
@@ -95,8 +135,12 @@ int main() {
     // Check for new version at startup
     if (conf::CHECK_LATEST_RELEASE) {
         const std::string latestVersion = fetchLatestVersion();
-        if (latestVersion.length() > 0 && latestVersion != conf::VERSION)
-            std::cout << "Update available! (" << latestVersion.substr(8) << ")\nSee https://wowtravis.com/mercury" << std::endl;
+        try {
+            if (latestVersion.length() > 0 && conf::isVersionOutdated(latestVersion))
+                std::cout << "Update available! (" << latestVersion.substr(8) << ")\nSee https://wowtravis.com/mercury" << std::endl;
+        } catch (std::invalid_argument&) {
+            std::cout << "Failed to compare local and remote versions!" << std::endl;
+        }
     }
 
     // Init server
@@ -140,6 +184,17 @@ int main() {
     std::vector<std::thread> threads;
     for (auto& server : serversVec)
         threads.emplace_back(std::thread([server]() { server->acceptLoop(); }));
+
+    // Wait for "exit" in cin (sets isExiting if "exit" is found)
+    while (!isExiting.load())
+        awaitExitCin();
+
+    /******* Reached if closing the program *******/
+    std::cout << "\nShutting down..." << std::endl;
+
+    // Kill the server
+    for (auto& server : serversVec)
+        server->kill();
 
     // Join all threads
     for (std::thread& t : threads)