+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ## v0.21.0
 - Add support for running Mercury outside the `bin` directory (#215)
     - Working across UNC paths for Windows as well
+- Added support for typing "exit" to exit (#221)
+    - Greatly improve thread closure handling
+
 
 ## v0.20.1
 - Fix Windows port checking if WinPHPCGIPath exists even if EnablePHPCGI is set to off (#217)